arm: Drop linker script when building without PIE
The current section headers generated by bionic are incompatible
with this (interp is in, ctors were dropped, and other tiny changes
related to Elf32_Dyn). But... we don't actually need this to
disable PIE, so drop it entirely instead of writing a new one.

ChaOS: Needed to boot on p930

Change-Id: I3b350a9f6251e60bf52114343d190f3d6d83b216